refactor: generator layer
