+++ src/entrypoints/generator.ts
@@ -1,1 +1,1 @@
-export { generateFiles as generate } from '../layers/4_generator/files.js'
+export { generate } from '../layers/4_generator/_.js'
