+++ src/layers/4_generator/_.ts
@@ -1,0 +1,2 @@
+export * from './config.js'
+export * from './generate.js'
