+++ src/layers/4_generator/generators/MethodsSelect.ts
@@ -1,13 +1,13 @@
 // todo jsdoc
 import { getNodeNameAndKind, isRootType } from '../../../lib/graphql.js'
-import { createModuleGenerator } from '../createCodeGenerator.js'
-import { renderName, title1 } from '../helpers.js'
-import { moduleNameSelectionSets } from './SelectionSets.js'
+import { createModuleGenerator } from '../helpers/moduleGenerator.js'
+import { renderName, title1 } from '../helpers/render.js'
+import { ModuleGeneratorSelectionSets } from './SelectionSets.js'
 
-export const { generate: generateMethodsSelect, moduleName: moduleNameMethodsSelect } = createModuleGenerator(
+export const ModuleGeneratorMethodsSelect = createModuleGenerator(
   `MethodsSelect`,
   ({ config, code }) => {
-    code.push(`import type * as $SelectionSets from './${moduleNameSelectionSets}.js'`)
+    code.push(`import type * as $SelectionSets from './${ModuleGeneratorSelectionSets.name}.js'`)
     code.push(`import type * as $Utilities from '${config.libraryPaths.utilitiesForGenerated}'`)
     code.push(``)
