+++ src/layers/4_generator/generators/SchemaBuildtime.ts
@@ -24,9 +24,10 @@ import {
   RootTypeName,
 } from '../../../lib/graphql.js'
 import { entries, values } from '../../../lib/prelude.js'
-import { createModuleGenerator } from '../createCodeGenerator.js'
-import { type Config } from '../generateCode.js'
-import { getDocumentation, getInterfaceImplementors } from '../helpers.js'
+import type { Config } from '../config.js'
+import { createModuleGenerator } from '../helpers/moduleGenerator.js'
+import { getDocumentation, getInterfaceImplementors } from '../helpers/render.js'
+import { ModuleGeneratorScalar } from './Scalar.js'
 
 const namespaceNames = {
   GraphQLEnumType: `Enum`,
@@ -266,11 +267,11 @@ const renderArg = (config: Config, arg: GraphQLArgument) => {
 
 // high level
 
-export const { generate: generateSchemaBuildtime, moduleName: moduleNameSchemaBuildtime } = createModuleGenerator(
+export const ModuleGeneratorSchemaBuildtime = createModuleGenerator(
   `SchemaBuildtime`,
   ({ config, code }) => {
     code.push(`import type * as $ from '${config.libraryPaths.schema}'`)
-    code.push(`import type * as $Scalar from './Scalar.ts'`)
+    code.push(`import type * as $Scalar from './${ModuleGeneratorScalar.name}.js'`)
     code.push(`\n\n`)
 
     for (const [name, types] of entries(config.typeMapByKind)) {
